refactor: move collections and services to data layer
