+++ src/app/api/auth/google/callback/route.test.ts
@@ -28,7 +28,7 @@ vi.mock('@/business-layer/security/google', async () => {
   }
 })
 
-vi.mock('@/collections/services/UserService', () => {
+vi.mock('@/data-layer/services/UserService', () => {
   return {
     default: vi.fn().mockImplementation(() => ({
       getUserByEmail: vi.fn().mockResolvedValue(userMock),
@@ -37,7 +37,7 @@ vi.mock('@/collections/services/UserService', () => {
   }
 })
 
-vi.mock('@/collections/services/AuthService', () => {
+vi.mock('@/data-layer/services/AuthService', () => {
   return {
     default: vi.fn().mockImplementation(() => ({
       createAuth: vi.fn().mockResolvedValue(authenticationMock),
